fix: optimize the Data convolution of license
