+++ src/main/java/org/opensourceway/sbom/manager/service/license/impl/LicenseServiceImpl.java
@@ -1,49 +1,25 @@
 package org.opensourceway.sbom.manager.service.license.impl;
 
 import com.github.packageurl.MalformedPackageURLException;
-import org.apache.commons.collections4.MapUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.tuple.Pair;
-import org.opensourceway.sbom.cache.constant.CacheConstants;
-import org.opensourceway.sbom.clients.license.LicenseClient;
-import org.opensourceway.sbom.clients.license.vo.ComplianceResponse;
-import org.opensourceway.sbom.clients.license.vo.LicenseInfo;
 import org.opensourceway.sbom.constants.BatchContextConstants;
 import org.opensourceway.sbom.constants.SbomConstants;
-import org.opensourceway.sbom.manager.dao.LicenseRepository;
-import org.opensourceway.sbom.manager.dao.PackageRepository;
-import org.opensourceway.sbom.manager.dao.ProductRepository;
 import org.opensourceway.sbom.manager.dao.RepoMetaRepository;
-import org.opensourceway.sbom.manager.model.ExternalPurlRef;
-import org.opensourceway.sbom.manager.model.License;
-import org.opensourceway.sbom.manager.model.Package;
 import org.opensourceway.sbom.manager.model.Product;
 import org.opensourceway.sbom.manager.model.RepoMeta;
 import org.opensourceway.sbom.manager.model.vo.PackageUrlVo;
 import org.opensourceway.sbom.manager.service.license.LicenseService;
 import org.opensourceway.sbom.manager.utils.PurlUtil;
-import org.opensourceway.sbom.manager.utils.cache.LicenseInfoMapCache;
-import org.opensourceway.sbom.manager.utils.cache.LicenseStandardMapCache;
 import org.opensourceway.sbom.openeuler.obs.SbomRepoConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
-import org.springframework.util.ObjectUtils;
 
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.UUID;
 
 @Service
 @Qualifier("LicenseServiceImpl")
@@ -51,51 +27,8 @@ import java.util.UUID;
 public class LicenseServiceImpl implements LicenseService {
     private static final Logger logger = LoggerFactory.getLogger(LicenseServiceImpl.class);
 
-    private static final Integer BULK_REQUEST_SIZE = 128;
-
-    @Autowired
-    private LicenseClient licenseClient;
-
-    @Autowired
-    private LicenseRepository licenseRepository;
-
-    @Autowired
-    private PackageRepository packageRepository;
-
-    @Autowired
-    private ProductRepository productRepository;
-
     @Autowired
     private RepoMetaRepository repoMetaRepository;
-
-    @Autowired
-    private LicenseInfoMapCache licenseInfoMapCache;
-
-    @Autowired
-    private LicenseStandardMapCache licenseStandardMapCache;
-
-    @Value("${isScan}")
-    private Boolean isScan;
-
-
-    private void scanLicense(ComplianceResponse element) {
-        try {
-            licenseClient.scanLicenseFromPurl(element.getPurl());
-        } catch (Exception e) {
-            logger.error("failed to scan license for purl {}", element.getPurl());
-            throw e;
-        }
-    }
-
-    private Boolean isContainPackage(Package pkg, License license) {
-        for (Package pkgs : license.getPackages()) {
-            if (pkgs.getId().equals(pkg.getId())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
 
     /***
      * change openEuler purl format to get license
@@ -162,59 +95,6 @@ public class LicenseServiceImpl implements LicenseService {
         }
     }
 
-    private void saveLicenseAndCopyrightForPackage(ComplianceResponse response, Package pkg) {
-        if (response.getResult().getRepoCopyrightLegal().size() != 0) {
-            pkg.setCopyright(response.getResult().getRepoCopyrightLegal().get(0));
-        }
-        if (response.getResult().getRepoLicense().size() != 0) {
-            pkg.setLicenseConcluded(response.getResult().getRepoLicense().get(0));
-        }
-    }
-
-    private License generateNewLicense(String lic) {
-        License license = new License();
-        license.setSpdxLicenseId(lic);
-
-        // use cache
-        Map<String, LicenseInfo> licenseInfoMap = licenseInfoMapCache.getLicenseInfoMap(CacheConstants.LICENSE_INFO_MAP_CACHE_KEY_DEFAULT_VALUE);
-        if (!ObjectUtils.isEmpty(licenseInfoMap) && licenseInfoMap.containsKey(lic)) {
-            LicenseInfo licenseInfo = licenseInfoMap.get(lic);
-            license.setName(licenseInfo.getName());
-            license.setUrl(licenseInfo.getReference());
-        }
-        return license;
-    }
-
-    private void setLicenseAndPkgInfo(Map<String, List<String>> illegalLicenseInfo, List<Pair<Package, License>> dataToSave, ExternalPurlRef purlRef, ComplianceResponse response) {
-        List<String> illegalLicenseList = response.getResult().getRepoLicenseIllegal();
-        List<String> licenseList = new ArrayList<>(illegalLicenseList);
-        licenseList.addAll(response.getResult().getRepoLicenseLegal());
-        Package pkg = packageRepository.findById(purlRef.getPkg().getId()).orElseThrow();
-        saveLicenseAndCopyrightForPackage(response, pkg);
-        licenseList.forEach(lic -> {
-            lic = licenseStandardMapCache.getLicenseStandardMap(CacheConstants.LICENSE_STANDARD_MAP_CACHE_KEY_PATTERN).getOrDefault(lic.toLowerCase(), lic);
-            License license = licenseRepository.findBySpdxLicenseId(lic).orElse(generateNewLicense(lic));
-            if (pkg.getLicenses() == null) {
-                pkg.setLicenses(new HashSet<>());
-            }
-            if (license.getPackages() == null) {
-                license.setPackages(new HashSet<>());
-            } else if (!isContainPackage(pkg, license)) {
-                pkg.getLicenses().add(license);
-                license.getPackages().add(pkg);
-            }
-            if (illegalLicenseList.contains(lic)) {
-                license.setIsLegal(false);
-                List<String> licList = illegalLicenseInfo.getOrDefault(purlRef.getPkg().getName(), new ArrayList<>());
-                licList.add(lic);
-                illegalLicenseInfo.put(purlRef.getPkg().getName(), licList);
-            } else {
-                license.setIsLegal(true);
-            }
-            dataToSave.add(Pair.of(pkg, license));
-        });
-    }
-
     public String getPurlsForLicense(PackageUrlVo packageUrlVo, Product product) {
         String purl = "";
         String productType = String.valueOf(product.getAttribute().get(BatchContextConstants.BATCH_PRODUCT_TYPE_KEY));
@@ -232,87 +112,6 @@ public class LicenseServiceImpl implements LicenseService {
             return null;
         }
         return purl;
-    }
-
-    @Override
-    public Integer getBulkRequestSize() {
-        return BULK_REQUEST_SIZE;
-    }
-
-    @Override
-    public boolean needRequest() {
-        return licenseClient.needRequest();
-    }
-
-    @Override
-    public List<Pair<Package, License>> extractLicenseForPurlRefChunk(UUID sbomId, List<ExternalPurlRef> externalPurlChunk) {
-        logger.info("Start to extract License for sbom {}, chunk size:{}", sbomId, externalPurlChunk.size());
-        Set<Pair<ExternalPurlRef, Object>> resultSet = new HashSet<>();
-        Product product = productRepository.findBySbomId(sbomId);
-
-        try {
-            Set<String> repoPurlSet = new HashSet<>();
-            Map<String, String> pkgRepoPurlTrans = new HashMap<>();
-            externalPurlChunk.forEach(purlRef -> {
-                String purlForLicense = getPurlsForLicense(purlRef.getPurl(), product);
-                if (!Objects.isNull(purlForLicense)) {
-                    repoPurlSet.add(purlForLicense);
-                    pkgRepoPurlTrans.put(purlRef.getPurl().toString(), purlForLicense);
-                }
-            });
-            List<String> purls = new ArrayList<>(repoPurlSet);
-            ComplianceResponse[] responseArr = licenseClient.getComplianceResponse(purls);
-            if (Objects.isNull(responseArr) || responseArr.length == 0) {
-                return List.of();
-            }
-            externalPurlChunk.forEach(ref ->
-                    Arrays.stream(responseArr)
-                            .filter(response -> StringUtils.equals(pkgRepoPurlTrans.get(ref.getPurl().toString()), response.getPurl()))
-                            .forEach(licenseObj -> resultSet.add(Pair.of(ref, licenseObj))));
-
-        } catch (Exception e) {
-            logger.error("failed to extract License for sbom {}", sbomId);
-            throw new RuntimeException(e);
-        }
-        logger.info("End to extract license for sbom {}", sbomId);
-        return getLicenseAndPkgToSave(resultSet);
-    }
-
-
-    public List<Pair<Package, License>> getLicenseAndPkgToSave(Set<Pair<ExternalPurlRef, Object>> externalLicenseRefSet) {
-        Map<String, List<String>> illegalLicenseInfo = new HashMap<>();
-        List<Pair<Package, License>> dataToSave = new ArrayList<>();
-        int numOfNotScan = 0;
-        for (Pair<ExternalPurlRef, Object> externalLicenseRefPair : externalLicenseRefSet) {
-            ExternalPurlRef purlRef = externalLicenseRefPair.getLeft();
-            ComplianceResponse response = (ComplianceResponse) externalLicenseRefPair.getRight();
-            if (response.getResult().getIsSca().equals("true")) {
-                setLicenseAndPkgInfo(illegalLicenseInfo, dataToSave, purlRef, response);
-            } else {
-                if (Boolean.TRUE.equals(isScan) && response.getPurl().startsWith("pkg:git")) {
-                    scanLicense(response);
-                    numOfNotScan++;
-                }
-            }
-        }
-        logger.info("The num of package not scanned license: {}", numOfNotScan);
-        Map<String, List<String>> chunkIllegalLicenseInfo = new HashMap<>();
-        illegalLicenseInfo.forEach((pkgName, licList) -> {
-            List<String> templist = chunkIllegalLicenseInfo.getOrDefault(pkgName, new ArrayList<>());
-            templist.addAll(licList);
-            chunkIllegalLicenseInfo.put(pkgName, templist);
-        });
-        if (MapUtils.isNotEmpty(chunkIllegalLicenseInfo)) {
-            logger.warn("illegal licenses info in chunks:{}", illegalLicenseInfo);
-        }
-        return dataToSave;
-    }
-
-
-    @Override
-    public void persistExternalLicenseRefChunk(Pair<Set<Package>, Set<License>> licenseAndPkgListToSave) {
-        packageRepository.saveAll(licenseAndPkgListToSave.getLeft());
-        licenseRepository.saveAll(licenseAndPkgListToSave.getRight());
     }
 
 }
